requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
mysqlclient==2.2.6
PyJWT==2.10.0
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.2
urllib3==2.2.3
uuid6==2024.7.10
